src/requirements.txt
absl-py==0.1.13
annoy==1.11.4
astor==0.6.2
astroid==1.6.3
attrs==17.3.0
beautifulsoup4==4.6.0
bleach==1.5.0
bokeh==0.12.10
boto==2.48.0
boto3==1.5.25
botocore==1.8.39
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
docutils==0.14
en-core-web-lg==2.0.0
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
fasttext==0.9.1
Flask==1.0
ftfy==4.4.3
future==0.16.0
gast==0.2.0
gensim==3.5.0
grpcio==1.10.1
h5py==2.8.0
hdbscan==0.8.12
html5lib==0.9999999
hyperas==0.4
hyperopt==0.1
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
isort==4.3.4
itsdangerous==0.24
javalang==0.11.0
jedi==0.11.1
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.22.0
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.0
mccabe==0.6.1
mistune==0.8.3
msgpack==0.5.0
msgpack-numpy==0.4.1
msgpack-python==0.5.4
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nltk==3.4.5
nose==1.3.7
notebook==5.7.8
numba==0.37.0
numpy==1.14.5
pandas==0.23.0
pandocfilters==1.4.2
parso==0.1.1
pathlib==1.0.1
pexpect==4.3.1
pickleshare==0.7.4
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
py==1.5.2
pybind11==2.2.3
pycodestyle==2.4.0
Pygments==2.2.0
pylint==1.8.4
pymongo==3.6.0
pyparsing==2.2.0
pytest==3.3.1
python-crfsuite==0.9.5
python-dateutil==2.7.3
pytz==2018.3
PyYAML==5.1
pyzmq==16.0.3
qtconsole==4.3.1
regex==2017.4.5
requests==2.20.0
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.6
spacy==2.0.11
tabulate==0.8.2
tensorboard==1.9.0
tensorflow==1.12.2
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.2
toolz==0.9.0
tornado==4.5.2
tqdm==4.19.5
traitlets==4.3.2
ujson==1.35
umap-learn==0.2.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.3
widgetsnbextension==3.0.8
wrapt==1.10.11
yapf==0.22.0
